feat(project): доработка по ТЗ

+++ project/apps/post/src/app/post/post.service.ts
@@ -1,7 +1,7 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { PostRepository } from './post.repository';
 import { UpdatePostDto } from './dto/update-post.dto';
-import { POST_NOT_FOUND } from './post.constant';
+import { POST_NOT_FOUND, USER_NOT_AUTHORIZED } from './post.constant';
 import { PostEntityAdapter} from './post-entity.factory';
 import { CreateContentPostDtoType } from './dto';
 import { PostContentEntity } from './entity/post-content.entity';
@@ -17,7 +17,7 @@ export class PostService {
     private readonly postRepository: PostRepository,
   ) {}
 
-  public async createNewPost(dto: CreateContentPostDtoType): Promise<PostContentEntity> {
+  public async createNewPost(dto: CreateContentPostDtoType, userId: string): Promise<PostContentEntity> {
     const {type, tags, ...content} = dto;
     const tagsLowerUnique = transformTags(tags);
     const newPostDraft = new PostEntityAdapter[type]({
@@ -26,7 +26,7 @@ export class PostService {
       ...content,
       isPublished: true,
       isRepost: false,
-      authorId: '',
+      authorId: userId,
       originPostId: '',
       originAuthorId: '',
       comments: [],
@@ -44,10 +44,13 @@ export class PostService {
     return existPost;
   }
 
-  public async updatePostEntity(postId: string, dto: UpdatePostDto): Promise<PostContentEntity> {
+  public async updatePostEntity(postId: string,  userId: string, dto: UpdatePostDto,): Promise<PostContentEntity> {
     const existPost = await this.postRepository.findById(postId);
     if (!existPost) {
       throw new NotFoundException(POST_NOT_FOUND);
+    }
+    if ( existPost.authorId !== userId) {
+      throw new UnauthorizedException(USER_NOT_AUTHORIZED);
     }
     const updateEntity = new PostEntityAdapter[existPost.type]({
       ...existPost,
@@ -56,10 +59,13 @@ export class PostService {
     return await this.postRepository.updateById(updateEntity);
   }
 
-  public async deletePostEntity(postId: string): Promise<void> {
+  public async deletePostEntity(postId: string, userId: string): Promise<void> {
     const existPost = await this.postRepository.findById(postId);
     if (!existPost) {
       throw new NotFoundException(POST_NOT_FOUND);
+    }
+    if ( existPost.authorId !== userId) {
+      throw new UnauthorizedException(USER_NOT_AUTHORIZED);
     }
     this.postRepository.deleteById(postId);
   }
@@ -68,14 +74,20 @@ export class PostService {
      return await this.postRepository.findMany(filter);
   }
 
-  public async repostPost(postId: string): Promise<PostContentEntity> {
-    const {id: originPostId, authorId: originAuthorId, ...originPost} = await this.postRepository.findById(postId);
+  public async repostPost(postId: string, userId: string): Promise<PostContentEntity> {
+    const {
+      id: originPostId,
+      authorId: originAuthorId,
+      publishedAt,
+      ...originPost
+    } = await this.postRepository.findById(postId);
     const repostedPost = new PostEntityAdapter[originPost.type]({
       ...originPost,
-      authorId: '',
+      authorId: userId,
       isRepost: true,
       originPostId,
-      originAuthorId
+      originAuthorId,
+      publishedAt: new Date()
     });
     return this.postRepository.save(repostedPost);
   }
